move Strictness enum to darglint.strictness module
