+++ darglint/config.py
@@ -22,6 +22,7 @@ from typing import (  # noqa
 )
 
 from .docstring.base import DocstringStyle
+from .strictness import Strictness
 
 
 def get_logger():  # type: () -> Logger
@@ -41,50 +42,6 @@ POSSIBLE_CONFIG_FILENAMES = (
 )
 
 DEFAULT_DISABLED = {'DAR104'}
-
-
-class Strictness(Enum):
-    """The minimum strictness with which to apply checks.
-
-    Strictness does not describe whether or not a check
-    should be applied. Rather, if a check is done, strictness
-    describes how intense/strict/deep the check should be.
-
-    Each level here describes what is required of the
-    docstring at the given level of strictness.  For example,
-    SHORT_DESCRIPTION describes the situation where one-liners are
-    allowed, and sections are not required.
-
-    If the docstring being checked contains more than the
-    allowed amount below, then it is assumed that everything
-    must be checked.
-
-    """
-
-    # Allow a single-line description.
-    SHORT_DESCRIPTION = 1
-
-    # Allow a single-line description followed by a long
-    # description, but no sections.
-    LONG_DESCRIPTION = 2
-
-    # Require everything.
-    FULL_DESCRIPTION = 3
-
-    @classmethod
-    def from_string(cls, strictness):
-        strictness = strictness.lower().strip()
-        if strictness in {'short_description', 'short'}:
-            return cls.SHORT_DESCRIPTION
-        if strictness in {'long_description', 'long'}:
-            return cls.LONG_DESCRIPTION
-        if strictness in {'full_description', 'full'}:
-            return cls.FULL_DESCRIPTION
-
-        raise Exception(
-            'Unrecognized strictness amount "{}".  '.format(strictness) +
-            'Should be one of {"short", "long", "full"}'
-        )
 
 
 class AssertStyle(Enum):
@@ -133,7 +90,7 @@ class Configuration(object):
     def __init__(self, ignore, message_template, style, strictness,
                  ignore_regex=None, enable=[], indentation=4,
                  assert_style=AssertStyle.LOG, log_level=LogLevel.CRITICAL):
-        # type: (List[str], Optional[str], DocstringStyle, Strictness, Optional[str], List[str], int, AssertStyle) -> None  # noqa: E501
+        # type: (List[str], Optional[str], DocstringStyle, Strictness, Optional[str], List[str], int, AssertStyle, LogLevel) -> None  # noqa: E501
         """Initialize the configuration object.
 
         Args:
