+++ darglint/driver.py
@@ -18,9 +18,9 @@ from .config import (
     get_config,
     get_logger,
     LogLevel,
-    Strictness,
 )
 from .docstring.base import DocstringStyle
+from .strictness import Strictness
 import darglint.errors
 from darglint.error_report import ErrorReport
