+++ tests/test_docstring.py
@@ -13,7 +13,7 @@ from random import (
     shuffle,
 )
 
-from darglint.config import Strictness
+from darglint.strictness import Strictness
 from darglint.docstring.base import Sections
 from darglint.docstring.docstring import Docstring
